1-djb2.c

+++ 0x1A-hash_tables/1-djb2.c
@@ -10,9 +10,9 @@
 unsigned long int hash_djb2(const unsigned char *str)
 {
 unsigned long int hash;
-int c;
+unsigned long int c;
 hash = 5381;
 while ((c = *str++))
-hash = ((hash << 5) + hash) + c;/* hash * 33 + c */
+hash = ((hash << 5) + hash) +c;
 return (hash);
 }
